+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Default Command+N keybinding for SpawnNewInstance on macOS
 - Vi mode for copying text and opening links
+- `CopySelection` action which copies into selection buffer on Linux/BSD
 
 ### Changed